// src/types/types.ts
interface ProductImage {
  imageUrl: string;
  public_id: string;
  altText?: string;
  imageIndex?: number;

  size?: string;

  color?: string;
}
// ...
export interface specsDetails {
  id: number;
  specsDetails: string;
}
export interface IProductAdd {
  name: string;
  price: number;
  description: string;
  stock: number;
  discountPrice: number;
  posterImageUrl: string;
  posterImagePublicId: string;
  hoverImageUrl: string;
  hoverImagePublicId: string;
  productImages: ProductImage[];
  spacification: Array<{ specsDetails: string; _id: string }>;
  sizes?: string[];
  additionalInformation: AdditionalInformation[];
  categories: number[];
  subcategories: number[];
  Meta_Title: string;
  Canonical_Tag: string;
  Meta_Description: string;
  Images_Alt_Text: string;
  sale_counter?: string;
  filters?: any[];
}
// ...
export interface AdditionalInformation {
  key?: string;
  value?: string;
  colors?: string[];
  dimension?: string[];
}